+++ CHANGELOG.md
@@ -2,6 +2,7 @@ UNRELEASED
 ----------
 
 * [ [#2012](https://github.com/digitalfabrik/integreat-cms/issues/2012) ] Exclude archived pages when cloning a region
+* [ [#2093](https://github.com/digitalfabrik/integreat-cms/issues/2093) ] Fix recurring events endpoint
 
 
 2023.2.1